requirements.txt
bleach==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
dbus-python==1.2.18
distro==1.5.0
filelock==3.19.1
fsspec==2025.7.0
gpg==1.15.1
hf-xet==1.1.8
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
kaggle==1.7.4.5
libcomps==0.1.18
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
nftables==0.1
numpy==1.20.1
packaging==20.9
perf==0.1
pillow==11.3.0
protobuf==6.30.2
pycairo==1.20.1
PyGObject==3.40.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==8.0.4
PyYAML==5.4.1
requests==2.32.3
rpm==4.16.1.3
safetensors==0.6.2
selinux==3.5
sepolicy==3.5
setools==4.4.3
six==1.15.0
sympy==1.14.0
systemd-python==234
text-unidecode==1.3
timm==1.0.19
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.4.0
webencodings==0.5.1
